+++ Framework/SendbirdUIKit.xcframework/ios-arm64_x86_64-simulator/SendbirdUIKit.framework/Headers/SendbirdUIKit-Swift.h
@@ -4274,7 +4274,7 @@ SWIFT_CLASS("_TtC13SendbirdUIKit24SBUParentMessageInfoView")
 @property (nonatomic) CGRect bounds;
 @property (nonatomic) CGRect frame;
 - (nonnull instancetype)initWithReuseIdentifier:(NSString * _Nullable)reuseIdentifier OBJC_DESIGNATED_INITIALIZER;
-- (nullable instancetype)initWithCoder:(NSCoder * _Nonnull)coder SWIFT_UNAVAILABLE;
+- (nullable instancetype)initWithCoder:(NSCoder * _Nonnull)coder SWIFT_UNAVAILABLE_MSG("'init' has been renamed to 'SBUParentMessageInfoView(frame:)'");
 /// Calls the <code>userProfileTapHandler()</code> when the user profile is tapped.
 /// \param sender tapGestureRecognizer
 ///
@@ -9128,7 +9128,7 @@ SWIFT_CLASS("_TtC13SendbirdUIKit24SBUParentMessageInfoView")
 @property (nonatomic) CGRect bounds;
 @property (nonatomic) CGRect frame;
 - (nonnull instancetype)initWithReuseIdentifier:(NSString * _Nullable)reuseIdentifier OBJC_DESIGNATED_INITIALIZER;
-- (nullable instancetype)initWithCoder:(NSCoder * _Nonnull)coder SWIFT_UNAVAILABLE;
+- (nullable instancetype)initWithCoder:(NSCoder * _Nonnull)coder SWIFT_UNAVAILABLE_MSG("'init' has been renamed to 'SBUParentMessageInfoView(frame:)'");
 /// Calls the <code>userProfileTapHandler()</code> when the user profile is tapped.
 /// \param sender tapGestureRecognizer
 ///